chore(vm): move vm utils outside bwip
