+++ Cargo.toml
@@ -36,6 +36,7 @@ members = [
     "core/lib/utils",
     "core/lib/vlog",
     "core/lib/multivm",
+    "core/lib/vm_utils",
     "core/lib/web3_decl",
 
     # Test infrastructure
